feat(dao): implement interface for each dao
- preparing test on chess service
