+++ src/test/java/chess/dao/EventDaoTest.java
@@ -15,11 +15,11 @@ class EventDaoTest {
 
     private static final String TEST_TABLE = "event_test";
     private static final String SETUP_TEST_DB_SQL = String.format("INSERT INTO %s (game_id, type, description) "
-                    + "VALUES (1, 'MOVE', 'a2 a4'), (1, 'MOVE', 'a7 a5'), (2, 'MOVE', 'a2 a3')", TEST_TABLE);
+            + "VALUES (1, 'MOVE', 'a2 a4'), (1, 'MOVE', 'a7 a5'), (2, 'MOVE', 'a2 a3')", TEST_TABLE);
 
     private static final String CLEANSE_TEST_DB_SQL = String.format("TRUNCATE TABLE %s", TEST_TABLE);
 
-    private final EventDao dao = new EventDao(TEST_TABLE);
+    private final EventDao dao = new EventDaoImpl(TEST_TABLE);
 
     @BeforeEach
     void setUp() {
